requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
psycopg2-binary
django-jazzmin
django-cors-headers
gunicorn>=21.2 
openpyxl>=3.1
